requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Faker==40.1.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.4
